Update the files to add modules
