+++ src/index.js
@@ -1,11 +1,5 @@
 import './style.css';
-
-const scorecard = [
-  { name: 'Ali', score: '100' },
-  { name: 'Marcus', score: '96' },
-  { name: 'Broad', score: '89' },
-  { name: 'Paul', score: '88' },
-];
+import scorecard from './modules/scorelist';
 
 function renderLeaderboard() {
   const scoreboardElement = document.getElementById('scoreboard');
