+++ src/modules/scorelist.js
@@ -1,0 +1,8 @@
+const scorecard = [
+  { name: 'Ali', score: '100' },
+  { name: 'Marcus', score: '96' },
+  { name: 'Broad', score: '89' },
+  { name: 'Paul', score: '88' },
+];
+
+export default scorecard;
